pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "line-bot-sdk>=3.16.3",
    "openai>=1.74.0",
    "psycopg2-binary>=2.9.10",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.40",
]
